requirements.txt
appnope==0.1.2
backcall==0.2.0
boto3==1.24.12
botocore==1.27.12
cassandra-driver==3.25.0
click==8.0.1
cql==1.4.0
decorator==5.0.9
geomet==0.2.1.post1
ipython==7.24.1
ipython-genutils==0.2.0
jedi==0.18.0
jmespath==1.0.1
matplotlib-inline==0.1.2
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.18
ptyprocess==0.7.0
Pygments==2.9.0
python-dateutil==2.8.2
s3transfer==0.6.0
six==1.16.0
thrift==0.13.0
traitlets==5.0.5
urllib3==1.26.9
wcwidth==0.2.5
